requirements.txt
Flask==3.0.2
python-dateutil==2.8.2
google-generativeai==0.3.2
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
Flask-Session==0.6.0 